+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.6.8] - 2025. 04. 15
 
 * Fix focus issue when header is a TextField or TextFormField ( by @doonfrs )
+* Add RowWrapperScreen to navigation and home screen; update demo with new feature tile ( by @doonfrs )
 
 ## [1.6.7] - 2025. 04. 14